requirements.txt
streamlit
pandas
scikit-learn
nltk
torch
transformers
python-docx
PyPDF2